# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Хочу_То — Rental Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: getListings
      tags: [listings]
      summary: Get listings with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Listings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /listings/{id}:
    get:
      operationId: getListingById
      tags: [listings]
      summary: Get listing details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /listings/{id}/unavailable-dates:
    get:
      operationId: getListingUnavailableDates
      tags: [listings]
      summary: Get unavailable dates for a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unavailable date ranges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DateRange"
...
  /users/{id}/listings:
    get:
      operationId: getUserListings
      tags: [users]
      summary: Get user's own listings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listing"
...
  /listings/{id}/reviews:
    get:
      operationId: getListingReviews
      tags: [reviews]
      summary: Get reviews for a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [health]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeRequest"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /contact:
    post:
      operationId: submitContactForm
      tags: [health]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactFormRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        pricePerDay:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        regionId:
          type: integer
        regionName:
          type: string
        photos:
          type: array
          items:
            type: string
        ownerId:
          type: integer
        ownerName:
          type: string
        ownerAvatar:
          type: string
        ownerPhone:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        deposit:
          type: number
        createdAt:
          type: string
      required:
        - id
        - title
        - pricePerDay
        - categoryId
        - regionId
        - ownerId
        - isAvailable
        - createdAt
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/Listing"
        - type: object
          properties:
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            ownerPhone:
              type: string
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - totalPages
    CreateListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        pricePerDay:
          type: number
        categoryId:
          type: integer
        regionId:
          type: integer
        photos:
          type: array
          items:
            type: string
        deposit:
          type: number
        isAvailable:
          type: boolean
      required:
        - title
        - pricePerDay
        - categoryId
        - regionId
    DateRange:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, confirmed]
      required:
        - startDate
        - endDate
...
    Review:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        text:
          type: string
        createdAt:
          type: string
      required:
        - id
        - listingId
        - authorId
        - authorName
        - rating
        - createdAt
...
    NewsletterSubscribeRequest:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    ContactFormRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string
        subject:
          type: string
      required:
        - name
        - email
        - message